+++ src/decorators/MessageExtensionDeclaration.ts
@@ -1,6 +1,10 @@
 // Copyright (c) Wictor Wilén. All rights reserved.
 // Licensed under the MIT license.
 
+/**
+ * Decorator function for Messaging Extensions
+ * @param name Name of the messaging extension
+ */
 export function MessageExtensionDeclaration(name: string): PropertyDecorator {
 
     return (target: any, propertyKey: string | symbol): void => {