+++ src/impl/index.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Wictor Wilén. All rights reserved.
-// Licensed under the MIT license.
-
-export { MessagingExtensionMiddleware } from "./MessagingExtensionMiddleware";