+++ src/index.ts
@@ -4,4 +4,3 @@
 export * from "./decorators";
 export * from "./interfaces";
 export * from "./routers";
-export * from "./impl";