+++ src/interfaces/index.ts
@@ -4,4 +4,3 @@
 export { IBot } from "./IBot";
 export { IConnector } from "./IConnector";
 export { IOutgoingWebhook } from "./IOutgoingWebhook";
-export { IMessageExtension } from "./IMessageExtension";