+++ src/routers/index.ts
@@ -1,5 +1,5 @@
 // Copyright (c) Wictor Wilén. All rights reserved.
 // Licensed under the MIT license.
 
-export { MsTeamsApiRouter } from "./MsTeamsApiRouter";
+export { default as MsTeamsApiRouter } from "./MsTeamsApiRouter";
 export { MsTeamsPageRouter } from "./MsTeamsPageRouter";